+++ cucumber-expressions/ruby/lib/cucumber/cucumber_expressions/parameter_type.rb
@@ -38,7 +38,7 @@ module Cucumber
         if @transformer.arity == 1
           non_nil_group_values = group_values.compact
           raise CucumberExpressionError.new(
-              "Single transformer unexpectedly matched 2 values - \"#{non_nil_group_values[0]}\" and \"${non_nil_group_values[1]}\""
+              "Single transformer unexpectedly matched 2 values - \"#{non_nil_group_values[0]}\" and \"#{non_nil_group_values[1]}\""
           ) if non_nil_group_values.length >= 2
           args = [non_nil_group_values[0]]
         else